Move components used by both login and group forms to forms-common
