+++ h/static/scripts/forms-common/components/Star.tsx
